// api.w9.io/utils/enum.ts
export enum ResponseStatusCode {
  Ok = 200,
  MultiStatus = 300,
  BadRequest = 400,
  Unauthorized = 401,
  CreditsNotSufficient = 402,
  Forbidden = 403,
  MethodNotAllowed = 405,
  NotFound = 404,
  UnsupportedMediaType = 415,
  RequestLimitExceeded = 429,
  InternalServerError = 500,
  NotAvailable = 503,
  InvalidDomain = 1001,
  RequiredDomain = 1002,
  PartnerUserAccessToken = 1003,
  InvalidIp = 1004,
}